# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Accounting app backend API for auth, categories, transactions, stats, and API keys.",
        "title": "Sumi API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/stats/category": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stats"
                ],
                "summary": "Get category stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Month in YYYY-MM",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Transaction type: 1 expense, 2 income",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.CategoryStatsOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stats/daily": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stats"
                ],
                "summary": "Get daily stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Month in YYYY-MM",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.DailyStatsOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stats/home": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stats"
                ],
                "summary": "Get monthly stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Month in YYYY-MM",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.MonthlyStatsOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stats/monthly": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stats"
                ],
                "summary": "Get monthly stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Month in YYYY-MM",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.MonthlyStatsOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "services.CategoryStatsItem": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer"
                },
                "category_name": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "parent_category_id": {
                    "type": "integer"
                },
                "parent_category_name": {
                    "type": "string"
                }
            }
        },
        "services.CategoryStatsOutput": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.CategoryStatsItem"
                    }
                },
                "month": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "services.DailyStatsDay": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.DailyStatsItem"
                    }
                }
            }
        },
        "services.DailyStatsItem": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "expense": {
                    "type": "string"
                },
                "income": {
                    "type": "string"
                }
            }
        },
        "services.DailyStatsOutput": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.DailyStatsDay"
                    }
                },
                "month": {
                    "type": "string"
                }
            }
        },
        "services.MonthlyStatsItem": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "net_amount": {
                    "type": "string"
                },
                "total_expense": {
                    "type": "string"
                },
                "total_income": {
                    "type": "string"
                }
            }
        },
        "services.MonthlyStatsOutput": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.MonthlyStatsItem"
                    }
                },
                "month": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Programmatic API key for server-to-server access.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Use the format: Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}